remove memory item lookup as too complex to handle item/work lookup case
