+++ requests/src/main/scala/weco/api/requests/models/RequestedItemWithWork.scala
@@ -4,4 +4,4 @@ import weco.catalogue.internal_model.identifiers.CanonicalId
 import weco.catalogue.internal_model.identifiers.IdState.Identified
 import weco.catalogue.internal_model.work.Item
 
-case class SourceIdentifierItemLookup(workId: CanonicalId, title: Option[String], item: Item[Identified])
+case class RequestedItemWithWork(workId: CanonicalId, title: Option[String], item: Item[Identified])
