--- requests/src/main/scala/weco/api/requests/services/ElasticItemLookup.scala
@@ -1,128 +1,0 @@
-package weco.api.requests.services
-
-import com.sksamuel.elastic4s.ElasticApi.fieldSort
-import com.sksamuel.elastic4s.ElasticDsl.{boolQuery, search, termQuery}
-import com.sksamuel.elastic4s.requests.searches.MultiSearchRequest
-import com.sksamuel.elastic4s.requests.searches.sort.SortOrder
-import com.sksamuel.elastic4s.{ElasticClient, Index}
-import grizzled.slf4j.Logging
-import weco.api.requests.models.SourceIdentifierItemLookup
-import weco.api.search.elasticsearch.{DocumentNotFoundError, ElasticsearchError, ElasticsearchService}
-import weco.catalogue.internal_model.Implicits._
-import weco.catalogue.internal_model.identifiers.{CanonicalId, IdState, SourceIdentifier}
-import weco.catalogue.internal_model.work.WorkState.Indexed
-import weco.catalogue.internal_model.work.{Item, Work}
-
-import scala.concurrent.{ExecutionContext, Future}
-
-class ElasticItemLookup(
-  elasticsearchService: ElasticsearchService,
-  index: Index
-)(
-  implicit ec: ExecutionContext
-) extends ItemLookup with Logging {
-
-  /** Returns the SourceIdentifier of the item that corresponds to this
-    * canonical ID.
-    *
-    */
-  def byCanonicalId(
-    itemId: CanonicalId
-  ): Future[Either[ElasticsearchError, Item[IdState.Identified]]] = {
-    val searchRequest =
-      search(index)
-        .query(
-          boolQuery.filter(
-            termQuery("data.items.id.canonicalId", itemId.underlying),
-            termQuery(field = "type", value = "Visible")
-          )
-        )
-        .size(1)
-
-    elasticsearchService.findBySearch[Work[Indexed]](searchRequest).map {
-      case Left(err) => Left(err)
-      case Right(works) =>
-        val item =
-          works
-            .flatMap { _.data.items }
-            .collectFirst {
-              case item @ Item(IdState.Identified(id, _, _), _, _, _)
-                  if id == itemId =>
-                // This .asInstanceOf[] is a no-op to help the compiler see what
-                // we can see by reading the code.
-                item.asInstanceOf[Item[IdState.Identified]]
-            }
-
-        item match {
-          case Some(it) => Right(it)
-          case None     => Left(DocumentNotFoundError(itemId))
-        }
-    }
-  }
-
-  override def bySourceIdentifier(
-    sourceIdentifiers: Seq[SourceIdentifier]
-  ): Future[Seq[Either[ElasticsearchError, SourceIdentifierItemLookup]]] = {
-    val multiSearchRequest = MultiSearchRequest(
-      sourceIdentifiers.map { sourceIdentifier =>
-        search(index)
-          .query(
-            boolQuery
-              .filter(termQuery(field = "type", value = "Visible"))
-              .should(
-                termQuery(
-                  "data.items.id.sourceIdentifier.value",
-                  sourceIdentifier.value
-                )
-              )
-          )
-          .sortBy(
-            fieldSort("state.sourceIdentifier.value").order(SortOrder.Asc)
-          )
-      }
-    )
-
-    elasticsearchService
-      .findByMultiSearch[Work[Indexed]](multiSearchRequest)
-      .map {
-        _.zip(sourceIdentifiers).map {
-          case (Right(works), srcId) =>
-            works
-              .flatMap { work =>
-                work.data.items.map(item => (work.data.title, work.state.canonicalId, item))
-              }
-              .collect {
-                case (title, workId, item @ Item(id @ IdState.Identified(_, _, _), _, _, _))
-                    if id.sourceIdentifier == srcId =>
-                  // This .asInstanceOf[] is a no-op to help the compiler see what
-                  // we can see by reading the code.
-                  item.asInstanceOf[Item[IdState.Identified]]
-                  SourceIdentifierItemLookup(
-                    workId = workId,
-                    title = title,
-                    item = item.asInstanceOf[Item[IdState.Identified]]
-                  )
-              }
-              .toList match {
-              // We can return multiple items from multiple works
-              // The sortBy above returns us the lowest bibId work association
-              case List(item) => Right(item)
-              case item :: _  => Right(item)
-              case List()     => Left(DocumentNotFoundError(srcId))
-            }
-
-          case (Left(err), _) => Left(err)
-        }
-      }
-  }
-}
-
-object ElasticItemLookup {
-  def apply(elasticClient: ElasticClient, index: Index)(
-    implicit ec: ExecutionContext
-  ): ElasticItemLookup =
-    new ElasticItemLookup(
-      elasticsearchService = new ElasticsearchService(elasticClient),
-      index = index
-    )
-}
