+++ requests/src/main/scala/weco/api/requests/services/RequestsService.scala
@@ -15,7 +15,7 @@ import scala.concurrent.{ExecutionContext, Future}
 
 class RequestsService(
   sierraService: SierraRequestsService,
-  itemLookup: ItemLookup
+  itemLookup: ItemLookup,
 )(implicit executionContext: ExecutionContext)
     extends Logging {
