+++ requests/src/test/scala/weco/api/requests/fixtures/RequestsApiFixture.scala
@@ -1,6 +1,7 @@
 package weco.api.requests.fixtures
 
 import akka.http.scaladsl.model.{HttpRequest, HttpResponse}
+import com.sksamuel.elastic4s.{ElasticClient, Index}
 import weco.api.requests.RequestsApi
 import weco.api.requests.services.{ItemLookup, RequestsService}
 import weco.api.search.models.ApiConfig
@@ -21,13 +22,15 @@ trait RequestsApiFixture extends SierraServiceFixture {
     )
 
   def withRequestsApi[R](
-    itemLookup: ItemLookup,
-    responses: Seq[(HttpRequest, HttpResponse)] = Seq()
+                          elasticClient: ElasticClient,
+                          index: Index,
+                          responses: Seq[(HttpRequest, HttpResponse)] = Seq()
   )(testWith: TestWith[Unit, R]): R =
     withSierraService(responses) { sierraService =>
+
       val requestsService = new RequestsService(
         sierraService = sierraService,
-        itemLookup = itemLookup
+        itemLookup = ItemLookup(elasticClient, index = index)
       )
 
       val api: RequestsApi = new RequestsApi(requestsService)
